Improving convergence
HS2 is not yet solving and I think that it has to do with the numbers in the Jacobian. I tried a manual solution here. It seems better but it is not yet solving. What I did, is to look at the output of the Jacobian for one node, and identify the rows that had a number that was higher than 1E4. Then, I divided those rows (Jacobian and constraint itself) by 100. So you set a breakpoint at the end of the Jacobian calculation and run this code to identify the lines.
t1 = full(output);
t2 = t1(ic,:);
max(t2')
So far, it is a manual hack, which is not ideal. Instead, it might work to just divide all row in nv by 100, which would be nicer. But this approach is fine to test if it works etc.

It might make sense to do something similar to HS1.